rename and rearrange
